requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.2
django-debug-toolbar==5.2.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
setuptools==80.3.1
sqlparse==0.5.3
typing-extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
wheel==0.45.1
psycopg==3.1.12
psycopg-binary==3.1.12